Resolve error-prone and codestyle issues
